tests/wpt/web-platform-tests/html/canvas/element/path-objects/2d.path.isPointInpath.invalid.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>Canvas test: 2d.path.isPointInpath.invalid</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.path.isPointInpath.invalid</h1>
<p class="desc">Verify isPointInPath throws exceptions with invalid inputs.</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("Verify isPointInPath throws exceptions with invalid inputs.");
_addTest(function(canvas, ctx) {

canvas.width = 200;
canvas.height = 200;
path = new Path2D();
path.rect(0, 0, 100, 100);
path.rect(25, 25, 50, 50);
// Testing invalid enumeration isPointInPath (w/ and w/o Path object');
assert_throws_js(TypeError, function() { ctx.isPointInPath(path, 50, 50, 'gazonk'); });
assert_throws_js(TypeError, function() { ctx.isPointInPath(50, 50, 'gazonk'); });

// Testing invalid type isPointInPath with Path object');
assert_throws_js(TypeError, function() { ctx.isPointInPath(null, 50, 50); });
assert_throws_js(TypeError, function() { ctx.isPointInPath(null, 50, 50, 'nonzero'); });
assert_throws_js(TypeError, function() { ctx.isPointInPath(null, 50, 50, 'evenodd'); });
assert_throws_js(TypeError, function() { ctx.isPointInPath(null, 50, 50, null); });
assert_throws_js(TypeError, function() { ctx.isPointInPath(path, 50, 50, null); });
assert_throws_js(TypeError, function() { ctx.isPointInPath(undefined, 50, 50); });
assert_throws_js(TypeError, function() { ctx.isPointInPath(undefined, 50, 50, 'nonzero'); });
assert_throws_js(TypeError, function() { ctx.isPointInPath(undefined, 50, 50, 'evenodd'); });
assert_throws_js(TypeError, function() { ctx.isPointInPath(undefined, 50, 50, undefined); });
assert_throws_js(TypeError, function() { ctx.isPointInPath([], 50, 50); });
assert_throws_js(TypeError, function() { ctx.isPointInPath([], 50, 50, 'nonzero'); });
assert_throws_js(TypeError, function() { ctx.isPointInPath([], 50, 50, 'evenodd'); });
assert_throws_js(TypeError, function() { ctx.isPointInPath({}, 50, 50); });
assert_throws_js(TypeError, function() { ctx.isPointInPath({}, 50, 50, 'nonzero'); });
assert_throws_js(TypeError, function() { ctx.isPointInPath({}, 50, 50, 'evenodd'); });


});
</script>
